# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Golang api of demo",
        "title": "Golang Esign API",
        "termsOfService": "http://github.com",
        "contact": {
            "name": "API Support",
            "url": "http://www.cnblogs.com",
            "email": "×××@qq.com"
        },
        "version": "1.0"
    },
    "host": "127.0.0.1:8081",
    "paths": {
        "/api/v1/release/template/add": {
            "post": {
                "description": "新增VueData",
                "tags": [
                    "release template"
                ],
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 31a165baebe6dec616b1f8f3207b4273",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "JSON数据",
                        "name": "vue",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.VueData"
                        }
                    }
                ],
                "responses": {
                    "20000": {
                        "schema": {
                            "$ref": "#/definitions/vuedata.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Platforms": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "gorm.Model",
                    "type": "string"
                },
                "vueDataID": {
                    "type": "integer"
                }
            }
        },
        "models.VueData": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "comment_disabled": {
                    "type": "boolean"
                },
                "content": {
                    "type": "string"
                },
                "content_short": {
                    "type": "string"
                },
                "display_time": {
                    "type": "string"
                },
                "forecast": {
                    "type": "number"
                },
                "image_uri": {
                    "type": "string"
                },
                "importance": {
                    "type": "integer"
                },
                "pageviews": {
                    "type": "integer"
                },
                "platforms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Platforms"
                    }
                },
                "reviewer": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "timestamp": {
                    "description": "gorm.Model",
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "vuedata.response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "object"
                }
            }
        }
    }
}